+++ _tests_/react-ts-tw/eslint.config.js
@@ -188,6 +188,7 @@ export default [ // eslint-disable-line
             "code-style/function-params-per-line": "error",
             "code-style/hook-callback-format": "error",
             "code-style/hook-deps-per-line": "error",
+            "code-style/use-state-naming-convention": "error",
             "code-style/if-else-spacing": "error",
             "code-style/if-statement-format": "error",
             "code-style/import-format": "error",